bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-decouple==3.8